Industrial map test
- Adjust footsteps
- Adjust alarm sound volume
- Change prefab variations to have 0.5 offset when spawned
- Fix prefab variation rotations
- Fix prefab models to be in origin
- Adjust prefab offset X and Y to be floats instead of ints
- Add three more maps: industrial, test with all objects, demo house
- Add factory model prefab, textures etc.
- Remove extra audio sources in XR Controller prefab causing alarm sound on startup

+++ Assets/Scripts/Map/MapObjects.cs
@@ -20,8 +20,8 @@ namespace MapObjects
         public string name;
         public GameObject prefab;
         public int rotation;
-        public int offsetX;
-        public int offsetY;
+        public float offsetX;
+        public float offsetY;
     }
 
     [Serializable]
